--- css/styles.css
@@ -1,20 +1,0 @@
-/* custom styles */
-
-.valign-wrapper {
-  height: 500px;
-}
-
-.container {
-  width: 600px;
-}
-
-.error {
-  background-color: #fff7e6;
-  height: 40px;
-  position: fixed;
-  width: 100%;
-}
-
-span.badge.new:after {
-    content: "";
-}
