+++ designs/error.html
@@ -94,4 +94,10 @@
         }
     </style>
 </head>
-<body>
+<body><div id="error">There was an error getting the profile for new. (Not Found)</div><img src="http://i.imgur.com/VKKm0pn.png" alt="Magnifying Glass" id="searchIcon">
+
+<form action="/" method="POST">
+    <input type="text" placeholder="Enter a Treehouse username" id="username" name="username">
+    <input type="submit" value="submit" class="button">
+</form></body>
+</html>
