+++ designs/index.html
@@ -94,4 +94,17 @@
         }
     </style>
 </head>
-<body>
+<body><img src="http://i.imgur.com/VKKm0pn.png" alt="Magnifying Glass" id="searchIcon">
+
+<form action="/" method="POST">
+    <input type="text" placeholder="Enter a Treehouse username" id="username" name="username">
+    <input type="submit" value="search" class="button">
+</form></body>
+</html>
+
+
+
+
+
+
+
