+++ designs/profile.html
@@ -94,4 +94,17 @@
         }
     </style>
 </head>
-<body>
+<body><div id="profile">
+
+    <img src="http://i.imgur.com/KMi72wU.png" alt="Avatar" id="avatar">
+
+    <p><span>@waldo</span></p>
+
+    <ul>
+        <li><span>154</span> Badges earned</li>
+        <li><span>652</span> JavaScript points</li>
+        <li><a href="/">search again</a></li>
+    </ul>
+
+</div></body>
+</html>
