--- index.html
@@ -1,36 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>###</title>
-  <meta name="description" content="###">
-  <meta name="###" content="###">
-  <link rel="stylesheet" href="css/materialize.css">
-  <link rel="stylesheet" href="css/styles.css">
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <!--[if lt IE 9]>
-  <script src="http://html5shiv.googlecode.com/svn/trunk/html5.js"></script>
-  <![endif]-->
-</head>
-
-<body>
-  <div class="valign-wrapper">
-    <div class="container center-align">
-      <div class="row center-align"> <i class="large material-icons">search</i> </div>
-      <div class="row">
-        <form action="/" method="POST">
-          <div class="row">
-            <label for="first_name">
-              <h5>Profile UserName</h5></label>
-            <input placeholder="User Name" id="first_name" type="text" class="validate"> </div>
-          <div class="row"> <a href="" class="waves-effect waves-light btn">SEARCH PROFILE</a> </div>
-        </form>
-      </div>
-    </div>
-  </div>
-</body>
-
-</html>
