--- npm-debug.log
@@ -1,21 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ 'C:\\Program Files (x86)\\nodejs\\node.exe',
-1 verbose cli   'C:\\Program Files (x86)\\nodejs\\node_modules\\npm\\bin\\npm-cli.js',
-1 verbose cli   'start' ]
-2 info using npm@3.3.12
-3 info using node@v5.2.0
-4 verbose stack Error: ENOENT: no such file or directory, open 'C:\Users\micposso\Desktop\NODE\treehouse\node-profile-search\package.json'
-4 verbose stack     at Error (native)
-5 verbose cwd C:\Users\micposso\Desktop\NODE\treehouse\node-profile-search
-6 error Windows_NT 10.0.14393
-7 error argv "C:\\Program Files (x86)\\nodejs\\node.exe" "C:\\Program Files (x86)\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" "start"
-8 error node v5.2.0
-9 error npm  v3.3.12
-10 error path C:\Users\micposso\Desktop\NODE\treehouse\node-profile-search\package.json
-11 error code ENOENT
-12 error errno -4058
-13 error syscall open
-14 error enoent ENOENT: no such file or directory, open 'C:\Users\micposso\Desktop\NODE\treehouse\node-profile-search\package.json'
-14 error enoent This is most likely not a problem with npm itself
-14 error enoent and is related to npm not being able to find a file.
-15 verbose exit [ -4058, true ]
