+++ renderer.js
@@ -1,0 +1,22 @@
+var fs = require("fs");
+
+function mergeValues(values, content) {
+  //Cycle over the keys
+  for(var key in values) {
+    //Replace all {{key}} with the value from the values object
+    content = content.replace("{{" + key + "}}", values[key]);
+  }
+  //return merged content
+  return content;
+}
+
+function view(templateName, values, reponse) {
+  //Read from the template file
+  var fileContents = fs.readFileSync('./views/' + templateName + '.html', {encoding: "utf8"});
+  //Insert values in to the content
+  fileContents = mergeValues(values, fileContents);
+  //Write out the contents to the response
+  reponse.write(fileContents);
+}
+
+module.exports.view = view;
