+++ views/error.html
@@ -1,4 +1,1 @@
-  <div class="row error">
-    <div class="col s12 center-align">
-      <h6>There was an error handeling the profile for new. (Not Found)</h6></div>
-  </div>
+<div id="error">{{errorMessage}}</div>
