+++ views/search.html
@@ -1,12 +1,6 @@
-<div class="container center-align">
-  <div class="row center-align"> <i class="large material-icons">search</i> </div>
-  <div class="row">
-    <form action="/" method="POST">
-      <div class="row">
-        <label for="first_name">
-          <h5>Profile UserName</h5></label>
-        <input placeholder="User Name" id="first_name" type="text" class="validate"> </div>
-      <div class="row"> <a href="" class="waves-effect waves-light btn">SEARCH PROFILE</a> </div>
-    </form>
-  </div>
-</div>
+<img src="http://i.imgur.com/VKKm0pn.png" alt="Magnifying Glass" id="searchIcon">
+
+<form action="/" method="POST">
+    <input type="text" placeholder="Enter a Treehouse username" id="username" name="username">
+    <input type="submit" value="search" class="button">
+</form>
